more int to float

+++ src/main/java/ichttt/mods/firstaid/api/damagesystem/AbstractPlayerDamageModel.java
@@ -110,7 +110,7 @@ public abstract class AbstractPlayerDamageModel implements Iterable<AbstractDama
 
     public abstract void setAbsorption(float absorption);
 
-    public abstract int getCurrentMaxHealth();
+    public abstract float getCurrentMaxHealth();
 
     @SideOnly(Side.CLIENT)
     public abstract int getMaxRenderSize();
